Updates to defunc and definec. I added a number of keyword arguments, including the following. 1. :typed-undef : this is set to t and it indicates whether the undef functions generated by defunc and definec are typed. For example if a function whose return type is :bool is defined, then the default setting will use a :bool type for the undef part of the definition, i.e., the function defined is constrained to always return a boolean. If a value of :nil is used, then the return type is :all. This is actually useful for CS2800, where I don't go into the details of definec/ defunc and it is useful in some cases where definitions depened on constrained functions, so the enumerators also depend of functions that assume certain values for these constrained functions, i.e., we don't have a good way of generating a witness, so making the undef case untyped is useful. Supporting this keyword required a number of changes to the core functions for defunc/definec.
2. :generalized-contract-thm : the default value is t and it indicates whether
the contract theorem generated should be generalized. What this means is that
we try to generate a contract theorem that is unconditional, even if there
are non-trivial input contracts. In some cases, we can get away with unconditional
contract theorems, especially given the note regarding :typed-undef above.
In some cases, that is not possible and so defunc/definec will revert to using
the hypotheses. If this is set to nil, then we do not try to generalize the
contract theorem. This is useful for CS2800, where we do not go into these
kinds of optimizatons. It is also useful if we want to carefully account for
the type hypotheses.
A number of core functions in defunc/definec were updated to suppport this.

I also update the debugging code to more generate more useful forms after
failures.