[Java] Add and parse an ABNF grammar of Java.
This consists of the complete lexical and syntactic (sub-)grammars from the Java
language specification.

The Java language specification uses its own grammar notation, which is actually
a bit closer to EBNF than ABNF. However, since we have a formalization and a
verified grammar parser for ABNF, but not for EBNF, we are content with using
ABNF to capture the Java grammar formally. ABNF and EBNF are not too different
anyways. The ABNF grammar of Java is as close to the grammar in the Java
language specification as possible: the correspondence should not be hard to
validate.

The verified ABNF grammar parser from the ABNF library has been successfully run
on this grammar of Java, obtaining a formal representation of the Java grammar
in ACL2, which can be used to formally specify the concrete syntax of Java at a
very high and assured level. As sanity checks, the parsed grammar is proved
(calcualated, in fact, since it's a constant) to be well-formed (according to
certain expected properties of valid ABNF grammars) and closed (i.e. every
nonterminal referenced on the right-hand side of some rule has a defining rule).

(There is no parser for Java here (yet). The grammar of Java itself, written in
ABNF, has been parsed using the ABNF grammar parser, i.e. a parser for the ABNF
notation, not for Java or other languages. However, parsing the grammar is the
first step towards specifying formal requirements for a parser of the Java
language.)