[Java] Improve ATJ implementation.
The translation from ACL2 variable names to Java variable names, which presents
some complexity, has been factored out of the functions that translate ACL2
terms to Java blocks and expressions. The variable name translation is now
realized as an ACL2-term-to-ACL2-term translation, where the new ACL2 terms have
variables whose names are all valid Java variable names, and where different
ACL2 variables with the same symbol (which is possible and common, due to scope
nesting) are suitably disambiguated.

This way, the subsequent translation phase, which goes from ACL2 to Java, can
simply use the SYMBOL-NAMEs of the ACL2 variables as the names of the
corresponding Java variables. This makes this translation phase simpler,
compared to the previous implementation. In particular, some arguments and
results of the ACL2 code generation functions that translate ACL2 to Java have
been eliminated.

The fact that the variable renaming is realized within ACL2 makes it easier to
verify its correctness. This variable renaming phase could be perhaps even
realized as an APT(-like) transformation that generates a proof: this would be
useful or necessary to generate an end-to-end proof of the code generation
process, if/when the ACL2-to-Java translation is verified (or verifying) as
well.

There may be other opportunities to factor some ACL2-to-ACL2 translations out of
the current ACL2-to-Java translation.