[Java] Improve ATJ.
Improve the generation of the wrappers of the natively implemented ACL2
functions to take into account the ATJ types of the functions directly, not just
whether the :GUARDS input is T or NIL. Also avoid including
types-for-natives.lisp into the tool. This allows potentially slightly different
choices of types for those ACL2 functions, and correspondingly slightly
different native implementations based on the types (e.g. having recognizers
such as STRINGP return a Java boolean instead of an Acl2Symbol object).