[Java] Add an ATJ type for the Java int type.
This is a step towards having ATJ generate Java code that manipulates Java
primitive values directly.

The newly added ATJ type corresponds to the INT-VALUE (fix)type in ACL2 (which
models Java int values in the language formalization) and to the int primitive
type in Java.

There is now also ATJ code to convert between this new type and the other Java
types.