[Std/system] Add a PURE-RAW-P system utility.
This says whether a function with raw Lisp code (i.e. a function satisfying
RAWP, another recently added system utility) is known to be free of side
effects, and thus whether its unnormalized body property accurately describes
its behavior (despite being implemented in raw Lisp for speed).

This is implemented by checking membership in a whitelist that can be easily
extended as needed.

The fact that a function with raw Lisp code is not in the whitelist does not
necessarily mean that the function has side effects; just that it is not known
(yet) to be side-effect-free.