Improved (one hopes) dependency analysis in books/projects/apply-model[-2]/Makefile.  Added missing include-books to support defxdoc forms in books/doc/relnotes.lisp.
Quoting :doc note-8-3-books:

  Modified custom makefiles books/projects/apply-model/Makefile and
  books/projects/apply-model-2/Makefile to avoid apparent mishandling
  of dependency analysis.

That makefile problem was causing numerous testing failures on leeroy,
but only recently.  I never have quite figured out the problem, though
it seems to have something to do with the inability of GNU Make 4.1
(on the linux system I use and on leeroy) to realize that "foo/../bar"
represents the same file as does "bar".  The new makefiles sacrifice a
bit of parallelism to reduce (I hope) the risk of such errors popping
up again.