[Java] Generalize an ATJ library extension.
The utility DECOMPOSE-AT-DOTS, which splits a string into the substrings
delimited by occurrences of the dot (.) character, has been generalized into a
new utility STRTOK!, which is very similar to STR::STRTOK, except that it does
not regard contiguous delimiters as one, and thus it may return empty
strings. For instance, (STRTOK! "abc.de..f") = ("abc" "de" "" "f").

This new function STRTOK! may be added to the Std/strings library, where STRTOK
already is. Another possibility could be, eventually, to merge the two utilities
into one that takes an argument flag specifying whether contiguous delimiters
should be regarded as one or not.