[Java] Improve ATJ.
Add a pre-translation step that removes all the unused lambda-bound variables,
i.e. the ones that do not occur free in the lambda body. The corresponding
actual arguments are removed too. This is safe and semantics-preserving, because
ATJ checks that all the code to be translated to Java is side-effect-free.