[Std/system] Add a new REMOVE-TRIVIAL-LAMBDA-VARS utility.
This removes formal parameters (and corresponding actual parameters) of lambda
expressions when formal and actual parameters are identical. In translated form,
ACL2 closes all the lambda expressions, but for certain term transformation
purposes it seems more convenient to work with possibly open lambda expressions,
where every formal parameter of the lambda expressions is a "true" let-bound
variable, and not an artifact of closure.

Also add some tests for this new utility.