[Java] Simplify ATJ implementation.
Add a pre-translation step that removes all the trivial lambda-bound variables,
i.e. the ones that are present just to close all the lambda expressions in ACL2
but that do not correspond to "true" LETs. After this step, we can treat LETs
(i.e. lambda applications) as terms that are not necessarily closed, as in other
languages (and also as in ACL2's untranslated terms), without having to optimize
away these "artificial" LET assignments at each pre-translation and translation
step. Overall, the ATJ code is shorter and simpler.