[Java] Improve ATJ pre-translation.
Swap the type annotation step and the variable renaming step so that the latter
comes after the former. This requires adapting the variable renaming step to
deal with the type annotations in variables.

This swap is motivated by upcoming extensions, in which we want to have type
annotations available when we rename variables.