[Java] Have ATJ generate some destructive updates.
Now ATJ performs an analysis to determine which LET-bound variables can be
reused in the generated Java code, so that instead of declaring different Java
local variables for different but homonymous ACL2 variables, the same Java local
variables are destructively updated when possible. For instance, a term like

(LET ((X ...))
  (LET ((X ...))
    ...))

is now translated to Java code of the form

<type> x = ...;
x = ...;
...

if the first value of X is not needed after assigning the second value to it.

This results in smaller and more readable code.

This is done via an additional ACL2-to-ACL2 pre-translation phase, in which ACL2
variables are marked as 'new' (meaning that new Java local variables must be
generated for them) or 'old' (meaning that an existing Java local variable can
be reused).

The variable renaming pre-translation step, which follows this new marking step,
had to be modified to generate new names for the 'new' variables while using
previous existing names for the 'old' variables. This necessitated the use of
two renaming maps, one for the 'new' and one for the 'old' variables, which must
be handled differently: the 'new' renaming map is passed through according to
the ACL2 scoping structure, while the 'old' renaming map has to be threaded
through consistently with the Java scoping structure that is used in the marking
step.

The ACL2-to-Java translation step has been modified to use the 'new' and 'old'
markings as "instructions" to generate either variable declarations with
initializers, or variable assignments.