[Java] Sealed the Java package of AIJ.
This means that, when the JVM is run, all the classes in AIJ's (Java) package
must come from the same JAR file that is sealed. This prevents access to
package-private members of AIJ by external code.

The JAR file is sealed both when generated by IntelliJ IDEA and when generated
by the Makefile.