[Java] Experiment improvement to code generated by ATJ.
Added an internal flag that determines whether, under suitable conditions, ATJ
generates Java conditional expressions a ? b : c instead of if-else statements
preceded by a new local variable that is assigned in the two branches. This is
for the shallow embedding.

While some generated Java methods look better this way, many look worse,
especially when the conditional expressions are nested. So for now this flag is
NIL; it may be removed eventually, but for now we keep it since the code
generator has that ability.