[Java] Change representation of index map in ATJ.
This is a simple but important change in representation. For the purpose of
renaming apart homonymous ACL2 variables that must become different Java
variablles, the keys of the index alist are now the Java variables instead of
the ACL2 variables. This way, the function that maps the ACL2 variables to the
Java variables (prior to adding indices) no longer needs to be injective,
therefore providing greater flexibility and wider options to turn ACL2 variables
into Java variables.