[Java] Simplify ATJ's handling of disallowed Java variable names.
We initialize the variable index map to pretend that Java keywords and the other
disallowed variable names have been already used, so that if we encounter such
variables an index will be always added to them, making them allowed. Thus, the
code no longer needs a special case to check for these variables explicitly.