[Java] Optimize generated Java code.
Turn every ACL2 quoted integer constant into a Java static final field that is
initialized once at class initialization time and then just referenced when
executing the Java methods generated from the ACL2 functions that contain the
quoted constants. This way, the ACL2 integer, which is a Java object, is not
re-created each time. This approach treats ACL2 integers as "literals" in Java.

A similar approach will be followed for ACL2 quoted constants of other types.

This change required moving the code to initialize the ACL2 environment to a
static initializer, because in general the ACL2 environment must be initialized
before constructing ACL2 values, and ACL2 values are constructed in the
initializers of the final static fields for the quoted integers.