[Std/typed-alists] Move to [books]/std.
From [books]/kestrel/std.

This was proposed on the ACL2 Books mailing list a few days ago. There were no
objections and one (private) expression of support.

This initially contains a few types of alists, but more can be easily added, as
needed. See the existing ones for a simple template.