[Java] Improve the structure of some ATJ code.
Now the Java methods for the ACL2 functions are calculated and stored into an
alist, which is then used to generate the classes.

This paves the way for generating additional members of the nested classes after
having processed all the functions/methods, as needed to generate static final
fields for quoted symbols in terms (this generation is upcoming).