[Java] Improve ATJ's tail recursion elimination.
To realize the parallel assignment to the method's parameters just before the
loop continuation statement, now we attempt a topological sort of the
assignments (using the dependency graph library at [books]/centaur/depgraph),
and we serialize the assignments according to the topological order if the
dependencies are not cyclical.

Otherwise, we revert to the simple strategy of introducing temporary variables
for all the parameters, initializing them with the expressions, and assigning
the temporary variables to the parameters. This could be improved, by reducing
the introduced temporary variables just as needed to break the dependency
cycles.