[Java] Generalize ATJ function tables.
Now the table supports, for each function, a 'main' function type and zero or
more 'other' function types. For now only the main one is used, but support for
the other ones will be added. The main function type is the one justified based
on the guards; the other function types will be narrower types that will be used
to generate more efficient overloaded variants of some methods.