[Java] Define a semilattice directly over ATJ Java types.
Operations for a meet semilattice are now defined directly over the Java types
generated by ATJ. This is isomorphic to the previously defined Java-based
semilattice on the ATJ types, but it provides a more clear separation between
the ATJ types, which really correspond to ACL2 types (i.e. predicates) and will
be eventually slightly renamed to reflect that, and the Java types that
correspond to such ACL2 types.

This separation opens the door to, and is in fact motivated by, more flexible
translations between ACL2 and Java types, in particular non-injective ones,
which are currently being considered.

The previous definition of the Java-based semilattice will be removed once the
code that uses it is redirected to use the new definition.