[Java] Improve AIJ's robustness.
Add checks requiring the first three ACL2 packages to be defined to be
"KEYWORD", "COMMON-LISP", and "ACL2" (in this order), the first and second ones
to import no symbols, and the third to import only symbols in the "COMMON-LISP"
package. This avoids malformed package definitions by exploiting the existence
of the pre-created symbols in the "COMMON-LISP" and "ACL2" packages, which are
created at class initialization time, before any package is actually defined.

Correspondingly add checks about the ACL2 packages in AIJ's assumptions file.