[Java] Improve ATJ's pretty-printer.
Now the pretty-printer introduces the necessary parentheses to preserve the tree
structure of the Java expressions, based on the operator priorities and
left/right associativity implicitly defined by the Java grammar.

The approach is based directly on the Java grammar, with ordered ranks
associated to expressions that are used to determine when parentheses are needed
or not: they are not needd when the actual rank is less than or equal to the
expected rank, otherwise the added parentheses effectively reduce the actual
rank so that if fits any expected rank. The approach is described in detail in
the documentation of the pretty-printer.