[Java] Improve code generated by ATJ.
Now that the pretty-printer automatically handles expression parenthesization,
we do not need to generate explicit parenthesized expressions, which we were
previously doing to ensure that the generated Java expressions are correctly
parsable, which was resulting in some extra parentheses being generated. Now the
code is slightly more readable and idiomatic.