[Java] Improve AIJ's variable index setting code.
Now the code that sets variable indices uniformly throws
IllegalArgumentException when the setting fails (no longer IllegalStateException
in some cases), because in essence the caller has supplied a malformed argument
as the definition of the function, at the top level.

The invariants about the map from variables to indices have been fleshed out and
documented.

The documentation of the conditions under which IllegalArgumentException is
thrown by the various involved methods has been made more clear and detailed.