[Java] Start adding support for multiple ATJ output types.
Now the DEF-ATJ-MAIN-FUNCTION-TYPE and DEF-ATJ-OTHER-FUNCION-TYPE macros
support, and in fact require, multiple output types for functions that return
multiple values. The rest of ATJ's code still treats all functions as returning
a single result, which is a CONSP in the case of a multi-result function. The
rest of ATJ's code will be extended to handle multiple types as needed.