[Java] Improve some ATJ code.
This relates to the ATJ function used to calculate the output type corresponding
to the minimum input types that are above or the same as the inferred argument
types of a function call. Now we always call that ATJ function with both primary
and secondary function types in all cases (before it was sometimes called only
on secondary types). The documentation has been improved to reflect this
increased uniformity, and documentation of the ATJ type annotation function that
call this ATJ function has been accordingly updated as well.