[Java] Improve generation of field expressions.
Not all expressions that read fields are grammatically field access expressions
in Java: the grammar rules for FieldAccess cannot generate expressions of the
form x.f where x and f are identifiers. Instead, this expression can be
generated from the rules for ExpressionName.

Thus, ATJ has been changed to generate expression names instead of field access
expressions when the left-hand side of the dot is an identifier (as opposed to a
different kind of expression). This eliminates an odd exception in the expected
ranks of field access expressions (exception with respect to the Java language
specification), which I had provisionally introduced so that x.f did not get
pretty-printed as (x).f.