[Java] Finish extending ATJ's type annotations for MV.
Now ATJ's type annotation pre-translation step fully supports lists of one or
more types, as opposed to just single types. A list of two or more types
corresponds to a multiple value, as produced by MV.

The ACL2-to-Java translation step still maps MV values to lists, but now that
the type annotations support MV, the next step is to extend the translation step
to use the generated MV classes instead.