[Java] Add an abstract syntax constructor.
This is for building Java expressions that access fields. According to the Java
grammar, expressions name.field where name is an expression name cannot be
derived from the FieldAccess nonterminal (from which all other field access
expressions can be derived); instead, it can be derived from ExpressionName (a
dot-separated sequence of identifiers). The distinction is important for
pretty-printing reasons, to avoid spurious parentheses as in (name).field. This
new function takes care of the distinction between the two kinds of
field-accessing expressions.