[Java] Extend some ATJ code for MV support.
Have the core translation functions handle non-empty lists of types instead of
single types. Only a few parts remain that coerce non-singleton lists of types
to the :ACONS type (so that MV values are treated as lists). Soon these will be
changed, as additional code to treat MV values specially will be added.