[Java] Improve ATJ's input validation.
Since generating tests for functions returning multiple values is not yet
supported after the recent change in which MV values are treated specially, now
ATJ stops with an informative error if one attempts to generate tests for
functions that return multiple values.

Also document this current restriction in the user documentation.