[Java] Improve the code generated by ATJ.
The pre-translation step that restores MV calls has been improved and extended
to restore MV calls not only at the top-level of functions, but also in
"internal" MV-LETs. This is not straightforward because MV-LETs and number of
results are not immediately apparent in translated terms (which is what ATJ
deals with), but the information can be recovered with a bit effort.

Now the Java code generated for the ABNF parser, which makes extensive use of
MV, no longer uses (the Java method corresponding to) MV-NTH, i.e. no objects
are allocated for the MV values created by the parser.

Compared to the Java code for the ABNF parser that was treating all the MV lists
as CONS lists, the new code appears to be roughtly 3x faster.