[Java] Extend the ATJ test generation code.
For functions that return multiple results, store the actual list of results in
the test structures. This is towards supporting test generation for multi-valued
functions. There is still an input validation check that functions used for test
generation must be single-valued. The check will be removed when other parts of
the code are extended appropriately.