[Java] Refactor and extend some ATJ code.
The ATJ code that generates the Java test methods (when a :TESTS input is
supplied) has been refactored so that the parts common to deep and shallow
embedding, specific to the deep embedding, and specific to the shallow
embedding, are more clearly factored and separated.

The code has also been generalized to handle tests with multiple outputs (for
ACL2 functions that return MV values). The input validation code still rejects
MV functions in :TESTS, but this limitation will be removed soon.