[APT] Streamline user documentation.
Remove the separate XDOC topics for the design notes of transformations, since
now the reference documentation pages describe the design notes hand in hand
with the various parts of the user documentation.

Moved the transformation topics one level up, directly under :DOC APT.

Explicitly ordered the subtopics of APT so that there are common concepts,
common options, then all the transformations, then the topic for the
(APT-specific) utilities.