[APT] Improve ISODATA.
Simplify the proof generated for the old-to-new theorem. This now makes use of
the new-to-old theorem, and is the same for all cases (recursive or not, and
predicate or not). This is all according to the design notes.

Also slighly rephrase the formula of the old-to-new theorem for the predicate
case, to make it consistent with the design notes and uniform with the
non-predicate case.