[Std/system] Add a utility to generate fresh logical names.
This is a better version of FRESH-NAME-IN-WORLD-WITH-$S in
[books]/kestrel/utilities/system/fresh-names.lisp, which will be removed after
all its uses are changed to the new utility.

The new utility uses FRESH-NAMEP-MSG-WEAK to make a more thorough freshness
check that also depends on the type of logical name; the type is a parameter of
the new utility. The new utility also uses just ADD-SUFFIX for logical names
that may be in the COMMON-LISP package, such as theorem names.