[APT] Improve ISODATA.
When the result of the target function is being transformed, now the application
of the FORTH isomorphism to the transformed body of the function descends into
any IF branches, resulting in a more idiomatic new function. That is, instead of
(FORTH (IF a b c)), now we generate (IF a (FORTH b) (FORTH c)) -- but if b
and/or c are IFs, we further push the FORTH into their branches.