[Java] Improve organization of language formalization.
Put the macros used to define ACL2 functions on Java primitive values into a
separate file, since they are used not just for the Java primitive operations,
but also for the Java primitive conversions.

Also generalize a bit the names of the macros, as well as the name of the
top-level XDOC topic for these macros.