Avoid ill-guarded calls of update-enabled-structure-array and clean up relevant code and comments.
This is fallout from a change made awhile ago but after Version 8.2
was released, so no release note is associated with this fix.

I noticed that for the ACL2 system function
update-enabled-structure-array, by adding the assertion (array1p name
(cons header alist)) which is the first conjunct of that function's
guard, then (include-book "centaur/sv/top" :dir :system) produced an
assertion violation.  I fixed that and a relevant comment.  While I
was doing that, I simplified the code by eliminating some complexity
that appears to have only negligible performance benefit.

I also added a comment describing some experiments that convinced me
that it's probably not in the cards to improve performance for this
function.  It seemed to be the last low-hanging fruit for speeding up
include-book, as it was taking about 5% to 6% of the total time for
including the book mentioned above.