[numbered names] Improve generation of next indexed name.
The code to generate the next indexed name was always returning a symbol in the
same package as the starting name. When that package is "COMMON-LISP",
attempting to use the next indexed name as a function name (as APT
transformations do), was causing a (not too informative) error. Now instead the
next numbered name is put into the "ACL2" package if the starting name is in the
"COMMON-LISP" package.