[APT] Simplify ISODATA.
Remove the :NORMALIZE option. The new function is always generated with the
default, i.e. :NORMALIZE T, just like DEFUN in ACL2.

Supporting :NORMALIZE in (all) APT transformations seems like an unnecessary
complication. The utility INSTALL-NOT-NORMALIZED can be easily used to "switch"
to using the non-normalized definition of a function, if needed.