[APT] Improve ISODATA user documentation.
Treat all arguments uniformly, without singling out the ones being transformed,
but instead introducing a notation for n isomorphic mappings (one for each
argument) and one for the optional isomorphic mapping for the result. This paves
the way to extending the transformation to handle multiple isomorphic mappings
for disjoint partitions of arguments and results.