[Java] Start adding JUnit tests to AIJ.
This commit adds a single simple test, but also adds the necessary libraries to
the IntelliJ IDEA project, and also a run configuration for the tests.

The test code is in a directory 'test' that is parallel to 'src', which seems
adequate since for now (and probably for the foreseeable future) AIJ will only
include Java code.

The AIJ makefile does not run the JUnit tests yet from the command line. These
tests run in IntelliJ IDEA.

The tests remain outside the AIJ jar, which only contains AIJ proper. The
compiled class files for the tests go into a separate directory from the class
files of AIJ proper. This is all handled by IntelliJ IDEA, via a few GUI setting
selections.

More JUnit tests are easy to add now.