[Java] Extend code generated by ATJ.
Generate eight Java methods to destructively update Java primitive arrays, one
for each Java primitive type.

This is just the first step in extending ATJ to generate Java code that
destructively updates arrays. These newly generated eight methods will be the
translation targets of the eight ACL2 functions that model Java primitive array
writes. An upcoming stobj-like analysis will ensure that all the Java arrays are
treated single-threadedly by the ACL2 functions translated to Java, so that it
will be sound to translate the eight ACL2 array write functions to these eight
methods.

In the planned initial approach, for uniformity, the Java code generated by ATJ
will still return the destructively updated arrays, even though this would be
unnecessary because the aforementioned stobj-like analysis will ensure the
naming consistency of arrays. The plan is then to investigate a Java-to-Java
post-translation approach to avoid returning explicitly these arrays. An
alternative will be to do that directly during (shallow) code generation, but
that complicates the ACL2-to-Java translation and the future proof story.
Nonetheless, both approaches are on the table at this point.