[APT] Generalize and refactor some ISODATA code.
Now the code that generates the body of the new function handles the case in
which results of a multi-value function are being transformed.

In the process of making this extension, merged two previously separate
implementationfunctions that were handling two cases, by localizing the
difference to just a small portion of the new implementation function.