[Java] Rearrange the ATJ array name information.
Associate output array names not just per function but per function type. This
is currently necessary because a function may have multiple types, some of which
include array types and some of which do not. Thus, it is safer and clearer to
associate distinct array name information for the distinct function types of the
same function. The array analysis will also be performed not once per function,
but once per function type of the function. This will often be just one type.