[Std/util] Improve DEFISO.
Now DEFISO accepts :HINTS inputs that are not keyword-value lists (where the
keywords indicate which applicability conditions the respective hints apply),
but that are hints to be used for all the applicability conditions.

Use the newer event macro input processing utility for that, which does not
require the list of applicability condition keywords, since other code already
checks that there are no extraneous keywords. Thus, remove some other code
related to calculating the list of keywords.