[Java] Improve the partial order of ATJ types.
Separate the types for Java primitive values and arrays from the types for the
ACL2 built-in values. The rationale is that we want to avoid any implicit
conversions between Java primitive values and arrays and their Java
representations of their ACL2 representations.

This means that the ATJ types no longer form a join semilattice, but they do if
we adjoin an extra top element (for which we use NIL). So we define an extended
partial order with the extra element, we define the semilattice join operation
on the extended partial order, we prove the partial order and lattice
properties, and then we define a join operations restricted to ATJ types but
that may return the extra element. We do the same for lists of ATJ types.

We adapt the type annotation pre-translation step of ATJ to stop with an error
message if joining types from two branches of an IF results in the adjoined top
element. This situation is a (deep) input validation failure: the ACL2 functions
supplied by the user do no satisfy all the constraints necessary to generate
Java code. This is only relevant in the shallow embedding when guards are
assumed: in this situation, ATJ can generate Java code that manipulates Java
primitive values and arrays, but we now prevent any implicit mixing of those
with the ACL2 value objects.

Upcoming extensions to the type annotation pre-translation step will also
prevent other forms of implicit conversions between Java primitive values and
arrays and ACL2 value objects.

This is also a step towards supporting the upcoming single-threaded array
analysis.