[event macros] Improve an XDOC constructor.
This is for the description of the :HINTS input of an event macro. The new
description includes the case in which this input is not a keyword-value list
that specifies different hints for different applicability conditions, but a
list of hints to use for all the applicability conditions.

This is consistent with the recent extension of the :HINTS input processing
utility and with the recent addition of the applicability condition utilities.